Python/requirements-light.txt
numpy>=1.24.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.15.0
transformers>=4.35.0
safetensors>=0.3.0
trimesh>=4.0.0
psutil>=5.9.0
mmap-ninja>=1.0.0
ray[default]>=2.7.0
zstandard>=0.21.0
lz4>=4.3.2
blosc2>=2.2.8
GPUtil>=1.4.0
pygltflib>=1.16.0
python-fbx>=2022.1.0
pymeshfix>=0.17.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
shapely>=2.0.0
scipy>=1.11.0
meshio>=5.3.0
fenics>=2021.2.0
sympy>=1.12.0
vtk>=9.3.0
PyQt6>=6.6.0
scikit-learn>=1.3.0
plotly>=5.18.0
joblib>=1.3.0
lru-cache-decorator>=1.2.0
fonttools>=4.42.0ou
typing-extensions>=4.7.0
tqdm>=4.65.0
pytorch3d>=0.7.0
chamferdist>=1.0.0
matplotlib>=3.7.0
wandb>=0.15.0
noise>=1.2.2
Pillow>=10.0.0
numpy-quaternion>=2022.4.3

# Cloud Integration
boto3>=1.40.16
azure-storage-blob>=12.26.0
google-cloud-storage>=3.3.0
paramiko>=4.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
cryptography>=45.0.6
